+++ src/index.jsx
@@ -1,30 +1,9 @@
-import React, { useState } from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom';
 
-import Title from './Title';
-import ClickMeButton from './ClickMeButton';
-import CountButtons from './CountButtons';
+import App from './App';
 
-
-function App() {
-  const [showNumber, setShowNumber] = useState(0);
-
-  const handleClickMe = (upNumber) => {
-    setShowNumber(showNumber + upNumber);
-  };
-
-  return (
-    <div>
-      <Title />
-      <ClickMeButton
-        showNumber={showNumber}
-        handleClickMe={handleClickMe}
-      />
-      <CountButtons
-        handleClickMe={handleClickMe}
-      />
-    </div>
-  );
-}
-
-ReactDOM.render(<App />, document.getElementById('app'));
+ReactDOM.render(
+  <App />,
+  document.getElementById('app'),
+);
